// payload-types.ts
/* tslint:disable */
/* eslint-disable */
/**
 * This file was automatically generated by Payload.
 * DO NOT modify it by hand. Instead, modify your source Payload config,
 * and re-run `payload generate:types` to regenerate this file.
 */

export interface Config {
  collections: {
    users: User
  }
  globals: {}
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "users".
 */
export interface User {
  id: string
  name: string
  email: string
  role: "user" | "admin"
  phone?: string | null
  address?: {
    street?: string | null
    city?: string | null
    zipCode?: string | null
  } | null
  preferences?: {
    newsletter?: boolean | null
    notifications?: boolean | null
  } | null
  updatedAt: string
  createdAt: string
  resetPasswordToken?: string | null
  resetPasswordExpiration?: string | null
  salt?: string | null
  hash?: string | null
  loginAttempts?: number | null
  lockUntil?: string | null
  password: string | null
}